Fix generating certificate. Add OG tags for a website and fix right display for a file robots.txt

+++ web/website/models.py
@@ -78,7 +78,4 @@ class WoodItem:
   #     return rounded_up_number + 10 - last_number
 
   def round_on_decimals(number):
-    return round(number, -1)
-
-  
-  
+    return int(round(number, -1))
